+++ src/ms3/cli.py
@@ -1,18 +1,6 @@
 # -*- coding: utf-8 -*-
 """
-This is a skeleton file that can serve as a starting point for a Python
-console script. To run this script uncomment the following lines in the
-[options.entry_points] section in setup.cfg:
-
-    console_scripts =
-         fibonacci = ms3.skeleton:run
-
-Then run `python setup.py install` which will install the command `fibonacci`
-inside your current environment.
-Besides console scripts, the header (i.e. until _logger...) of this file can
-also be used as template for Python modules.
-
-Note: This skeleton file can be safely removed if not needed!
+Command line interface for ms3.
 """
 
 import argparse, os
@@ -37,21 +25,21 @@ def check(args):
         'level': args.level,
         'file': log,
     }
-    if args.file is None:
-        p = Parse(dir=args.root_dir, index=['key', 'fname'], file_re=r'\.mscx$', labels_cfg=labels_cfg, logger_cfg=logger_cfg)
-    else:
-        p = Parse(paths=args.file, index=['key', 'fname'], labels_cfg=labels_cfg, logger_cfg=logger_cfg)
+    if args.regex is None:
+        args.regex = r'\.mscx$'
+    p = Parse(args.dir, paths=args.file, file_re=args.regex, exclude_re=args.exclude, recursive=args.nonrecursive,
+              index=['key', 'fnames'], labels_cfg=labels_cfg, logger_cfg=logger_cfg)
     if '.mscx' not in p.count_extensions():
-        p.logger.warning("No MSCX files found.")
+        p.logger.warning("No MSCX files to check.")
         return
     p.parse_mscx()
+    res = True
     if not args.scores_only:
         wrong = p.check_labels()
         if wrong is None:
             res = None
         if len(wrong) == 0:
             p.logger.info("No syntactical errors.")
-            res = True
         else:
             if not args.assertion:
                 p.logger.warning(f"The following labels don't match the regular expression:\n{wrong.to_string()}")
@@ -65,14 +53,14 @@ def compare(args):
     logger_cfg = {
         'level': args.level,
     }
-    if args.file is None:
-        p = Parse(dir=args.root_dir, key='compare', index='fname', file_re=r'\.mscx$', logger_cfg=logger_cfg)
-    else:
-        p = Parse(paths=args.file, key='compare', index='fname', logger_cfg=logger_cfg)
+    if args.regex is None:
+        args.regex = r'\.mscx$'
+    p = Parse(args.dir, paths=args.file, file_re=args.regex, exclude_re=args.exclude, recursive=args.nonrecursive,
+                  key='compare', index='fnames', logger_cfg=logger_cfg)
     if len(p._score_ids()) == 0:
         p.logger.warning(f"Your selection does not include any scores.")
         return
-    p.add_rel_dir(args.rel_dir, suffix=args.suffix, score_extensions=args.extensions, new_key='old')
+    p.add_rel_dir(args.annotations, suffix=args.suffix, score_extensions=args.extensions, new_key='old')
     p.parse_mscx()
     p.add_detached_annotations()
     p.compare_labels('old', store_with_suffix='_reviewed')
@@ -103,7 +91,8 @@ def extract(args):
         'file': args.logfile,
         'path': args.logpath,
     }
-    p = Parse(args.mscx_dir, file_re=args.file, exclude_re=args.exclude, recursive=args.nonrecursive, labels_cfg=labels_cfg,
+
+    p = Parse(args.dir, paths=args.file, file_re=args.regex, exclude_re=args.exclude, recursive=args.nonrecursive, labels_cfg=labels_cfg,
               logger_cfg=logger_cfg, simulate=args.test, ms=args.musescore)
     p.parse_mscx(simulate=args.test)
     p.store_lists(root_dir=args.out,
@@ -155,31 +144,37 @@ def convert(args):
 
 
 def repair(args):
-    pass
+    print(args.dir)
 
 
 def run():
-    """Entry point for console_scripts
-    """
-    parser = argparse.ArgumentParser(
-        prog='ms3',
-        formatter_class=argparse.RawDescriptionHelpFormatter,
-        description='''\
+    # reusable argument sets
+    input_args = argparse.ArgumentParser(add_help=False)
+    input_args.add_argument('-d', '--dir', metavar='DIR', nargs='+', type=check_dir,
+                                help='Folder(s) that will be scanned for input files. Defaults to current working directory if no individual files are passed via -f.')
+    input_args.add_argument('-n', '--nonrecursive', action='store_false',
+                                help="Don't scan folders recursively, i.e. parse only files in DIR.")
+    input_args.add_argument('-r', '--regex', metavar="REGEX", default=r'(\.mscx|\.mscz)$',
+                                help="Select only file names including this string or regular expression.")
+    input_args.add_argument('-e', '--exclude', metavar="regex", default=r'^(\.|_)',
+                                help="Any files or folders (and their subfolders) including this regex will be disregarded.")
+    input_args.add_argument('-f', '--file', metavar='PATHs', nargs='+',
+                                help='Add path(s) of individual file(s) to be checked.')
+    input_args.add_argument('-l', '--level', metavar='{c, e, w, i, d}', default='i',
+                                help="Choose how many log messages you want to see: c (none), e, w, i, d (maximum)")
+
+    # main argument parser
+    parser = argparse.ArgumentParser(formatter_class=argparse.RawDescriptionHelpFormatter, description='''\
 --------------------------
-| Welcome to ms3 Parsing |
+| Welcome to ms3 parsing |
 --------------------------
 
 The library offers you the following commands. Add the flag -h to one of them to learn about its parameters. 
 ''')
-    parser.add_argument('-l', '--level', metavar='LOG_LEVEL', default='i',
-                                help="Choose how many log messages you want to see: d (maximum), i, w, e, c (none)")
+    subparsers = parser.add_subparsers(help='The action that you want to perform.', dest='action')
 
-    subparsers = parser.add_subparsers(help='The action that you want to perform.')
-
-    extract_parser = subparsers.add_parser('extract', help="Extract selected information from MuseScore files and store it in TSV files.")
-    extract_parser.add_argument('mscx_dir', metavar='MSCX_DIR', nargs='?', type=check_dir,
-                                default=os.getcwd(),
-                                help='Folder that will be scanned for MuseScore files (.mscx).')
+    extract_parser = subparsers.add_parser('extract', help="Extract selected information from MuseScore files and store it in TSV files.",
+                                           parents=[input_args])
     extract_parser.add_argument('-N', '--notes', metavar='folder', help="Folder where to store TSV files with notes.")
     extract_parser.add_argument('-L', '--labels', metavar='folder', help="Folder where to store TSV files with annotation labels.")
     extract_parser.add_argument('-M', '--measures', metavar='folder', help="Folder where to store TSV files with measure information.")
@@ -192,16 +187,11 @@ The library offers you the following commands. Add the flag -h to one of them to
     extract_parser.add_argument('-o', '--out', metavar='ROOT_DIR', type=check_and_create,
                                 help="""Make all relative folder paths relative to ROOT_DIR rather than to MSCX_DIR.
 This setting has no effect on absolute folder paths.""")
-    extract_parser.add_argument('-f', '--file', metavar="regex", default=r'(\.mscx|\.mscz)$',
-                                help="Select only file names including this regular expression.")
-    extract_parser.add_argument('-e', '--exclude', metavar="regex", default=r'^(\.|_)',
-                                help="Any files or folders (and their subfolders) including this regex will be disregarded.")
     extract_parser.add_argument('-m', '--musescore', default='auto', help="""Command or path of MuseScore executable. Defaults to 'auto' (attempt to use standard path for your system).
 Other standard options are -m win, -m mac, and -m mscore (for Linux).""")
     extract_parser.add_argument('-t', '--test', action='store_true', help="No data is written to disk.")
     extract_parser.add_argument('-p', '--positioning', action='store_true', help="When extracting labels, include manually shifted position coordinates in order to restore them when re-inserting.")
-    extract_parser.add_argument('-r', '--raw', action='store_false', help="When extracting labels, leave chord symbols encoded instead of turning them into a single column of strings.")
-    extract_parser.add_argument('-n', '--nonrecursive', action='store_false', help="Don't scan folders recursively, i.e. parse only files in MSCX_DIR.")
+    extract_parser.add_argument('--raw', action='store_false', help="When extracting labels, leave chord symbols encoded instead of turning them into a single column of strings.")
     extract_parser.add_argument('-u', '--unfold', action='store_true', help="Unfold the repeats for all stored DataFrames.")
     extract_parser.add_argument('--logfile', metavar='file path or file name', help="""Either pass an absolute file path to store all logging data in that particular file
 or pass just a file name and the argument --logpath to create several log files of the same name in a replicated folder structure.
@@ -213,12 +203,7 @@ subdirectory; otherwise, an individual log file is automatically created for eac
 
 
     check_parser = subparsers.add_parser('check', help="""Parse MSCX files and look for errors.
-In particular, check DCML harmony labels for syntactic correctness.""")
-    check_parser.add_argument('root_dir', metavar='MSCX_DIR', nargs='?', type=check_dir,
-                              default=os.getcwd(),
-                              help="""Folder that will be scanned for MuseScore files (.mscx). Defaults to the current
-working directory except if you pass -f/--file.""")
-    check_parser.add_argument('-f', '--file', metavar='PATHs', nargs='+', help='Add path(s) of individual file(s) to be checked.')
+In particular, check DCML harmony labels for syntactic correctness.""", parents=[input_args])
     check_parser.add_argument('-s', '--scores_only', action='store_true',
                               help="Don't check DCML labels for syntactic correctness.")
     check_parser.add_argument('--assertion', action='store_true', help="If you pass this argument, an error will be thrown if there are any mistakes.")
@@ -227,18 +212,13 @@ working directory except if you pass -f/--file.""")
 
 
     compare_parser = subparsers.add_parser('compare',
-        help="For MSCX files for which annotation tables exist, create another MSCX file with a coloured label comparison.")
-    compare_parser.add_argument('root_dir', metavar='SCORE_DIR', nargs='?', type=check_dir,
-                              default=os.getcwd(),
-                              help="""Folder that will be scanned for score files. Defaults to the current
-    working directory except if you pass -f/--file.""")
-    compare_parser.add_argument('-f', '--file', metavar='PATHs', nargs='+',
-                              help='Add path(s) of individual file(s) to be compared.')
-    compare_parser.add_argument('-r', '--rel_dir', metavar='PATH', default='../harmonies',
+        help="For MSCX files for which annotation tables exist, create another MSCX file with a coloured label comparison.",
+        parents = [input_args])
+    compare_parser.add_argument('-a', '--annotations', metavar='PATH', default='../harmonies',
                                 help='Path relative to the score file(s) where to look for existing annotation tables.')
     compare_parser.add_argument('-s', '--suffix', metavar='SUFFIX', default='',
                                 help='If existing annotation tables have a particular suffix, pass this suffix.')
-    compare_parser.add_argument('-e', '--extensions', metavar='EXT', nargs='+',
+    compare_parser.add_argument('-x', '--extensions', metavar='EXT', nargs='+',
                                 help='If you only want to compare scores with particular extensions, pass these extensions.')
     compare_parser.set_defaults(func=compare)
 
@@ -249,7 +229,7 @@ working directory except if you pass -f/--file.""")
                                 help='Path to folder with files to convert; can be relative to the folder where the script is located.')
     convert_parser.add_argument('target', metavar='TARGET_DIR', type=check_and_create, default=os.getcwd(),
                                 help='Path to folder for converted files. Defaults to current working directory.')
-    convert_parser.add_argument('-e', '--extensions', nargs='+', default=['mscx', 'mscz'],
+    convert_parser.add_argument('-x', '--extensions', nargs='+', default=['mscx', 'mscz'],
                                 help="List, separated by spaces, the file extensions that you want to convert. Defaults to mscx mscz")
     convert_parser.add_argument('-f', '--format', default='mscx',
                                 help="You may choose one out of {png, svg, pdf, mscz, mscx, wav, mp3, flac, ogg, xml, mxl, mid}")
@@ -267,10 +247,13 @@ working directory except if you pass -f/--file.""")
     convert_parser.set_defaults(func=convert)
 
     repair_parser = subparsers.add_parser('repair',
-                                          help="Apply automatic repairs to your uncompressed MuseScore files.")
+                                          help="Apply automatic repairs to your uncompressed MuseScore files.",
+                                          parents=[input_args])
     repair_parser.set_defaults(func=repair)
 
     args = parser.parse_args()
+    if args.file is None and args.dir is None:
+        args.dir = os.getcwd()
     if 'func' in args:
         args.func(args)
     else: