+++ src/ms3/utils.py
@@ -30,6 +30,7 @@ DCML_REGEX = re.compile(r"""
     )
     (?P<pedalend>\])?
 )?
+(\|(?P<cadence>((HC|PAC|IAC|DC|EC|PC)(\..+?)?)))?
 (?P<phraseend>(\\\\|\}\{|\{|\}))?$
             """, re.VERBOSE)
 """:obj:`str`
@@ -52,6 +53,7 @@ DCML_DOUBLE_REGEX = re.compile(r"""
                                     )
                                     (?P<pedalend>\])?
                                   )?
+                                  (\|(?P<cadence>((HC|PAC|IAC|DC|EC|PC)(\..+?)?)))?
                                   (?P<phraseend>(\\\\|\}\{|\{|\})
                                   )?
                                  )
@@ -69,6 +71,7 @@ DCML_DOUBLE_REGEX = re.compile(r"""
                                     )
                                     (?P<pedalend2>\])?
                                   )?
+                                  (\|(?P<cadence2>((HC|PAC|IAC|DC|EC|PC)(\..+?)?)))?
                                   (?P<phraseend2>(\\\\|\}\{|\{|\})
                                   )?
                                  )?
@@ -622,7 +625,7 @@ def fifths2str(fifths, steps, inverted=False):
 
 
 def get_ms_version(mscx_file):
-    with open(mscx_file) as file:
+    with open(mscx_file, encoding='utf-8') as file:
         for i, l in enumerate(file):
             if i < 2:
                 pass