+++ CHANGELOG.rst
@@ -2,10 +2,24 @@
 Changelog
 =========
 
+Version 0.4.8
+=============
+
+* now reads DCML labels with cadence annotations
+* unified command-line interface file options and included ``-f file.json``
+* Parse got more options for creating DataFrame index levels
+* Parse.measures property for convenience
+* bug fixes for better GitHub workflows
+
 Version 0.4.7
 =============
 
-* More correct computation of chord tones
+* Labels can be attached to MuseScore's Roman Numeral Analysis (RNA) layer
+  * parameter `label_type=1` in both `Score.attach_labels()` and `Parse.attach_labels()`
+  * `Annotations.remove_initial_dots()` before inserting into the RNA layer
+  * `Annotations.add_initial_dots()` before inserting into the absolute chord layer
+* interpret all `#vii` in major contexts as `vii` when computing chord tones
+* code cosmetics and bug fixes
 
 Version 0.4.6
 =============